refactor: Account nr
Changed accountNr to Long to accomodate full-length account numbers

+++ src/main/java/edu/ntnu/idi/stud/team10/sparesti/controller/BankController.java
@@ -63,14 +63,14 @@ public class BankController {
   /**
    * Get account details for an account number.
    *
-   * @param accountNr (int) The account number
+   * @param accountNr (Long) The account number
    * @param token (Jwt) The JWT token
    * @return (ResponseEntity<AccountDto>) The account details
    */
   @GetMapping("/account/details/{accountNr}")
   @Operation(summary = "Get account details for an account number.")
   public ResponseEntity<AccountDto> getAccountDetails(
-      @PathVariable int accountNr, @AuthenticationPrincipal Jwt token) {
+      @PathVariable Long accountNr, @AuthenticationPrincipal Jwt token) {
     Long userId = token.getClaim(USER_ID_CLAIM);
     AccountDto accountDetails = bankService.getAccountDetails(accountNr, userId);
     return ResponseEntity.ok(accountDetails);
@@ -101,8 +101,8 @@ public class BankController {
   @PutMapping("/account/transfer")
   @Operation(summary = "Transfer money from one account to another")
   public ResponseEntity<String> transferMoney(
-      @RequestParam Integer fromAccountNr,
-      @RequestParam Integer toAccountNr,
+      @RequestParam Long fromAccountNr,
+      @RequestParam Long toAccountNr,
       @RequestParam double amount,
       @AuthenticationPrincipal Jwt token) {
     Long ownerId = token.getClaim(USER_ID_CLAIM);
@@ -113,14 +113,14 @@ public class BankController {
   /**
    * Gets a list of all transactions by a singular account number
    *
-   * @param accountNr (Integer) The accountNr
+   * @param accountNr (Long) The accountNr
    * @return (ResponseEntity&lt;Set&lt;TransactionDto&gt; &gt;) Set of all transactions by the
    *     account.
    */
   @GetMapping("/transactions/{accountNr}")
   @Operation(summary = "Get all transactions by an account number")
   public ResponseEntity<Set<TransactionDto>> getAllTransactionsByAccountNr(
-      @PathVariable Integer accountNr, @AuthenticationPrincipal Jwt token) {
+      @PathVariable Long accountNr, @AuthenticationPrincipal Jwt token) {
     Long userId = token.getClaim(USER_ID_CLAIM);
     return ResponseEntity.ok().body(bankService.getTransactionsByAccountNr(accountNr, userId));
   }
@@ -128,7 +128,7 @@ public class BankController {
   /**
    * Get the last 30 days of transactions from an account, using its number.
    *
-   * @param accountNr (Integer) the accounts unique number in the bank.
+   * @param accountNr (Long) the accounts unique number in the bank.
    * @param token (Jwt) the JWT token
    * @return (ResponseEntity&lt;Set&lt;TransactionDto&gt;&gt;): a set including the last 30 days of
    *     transactions for the account.
@@ -136,7 +136,7 @@ public class BankController {
   @GetMapping("/transactions/recent/{accountNr}")
   @Operation(summary = "Get last 30 days of transactions from an account number")
   public ResponseEntity<Set<TransactionDto>> getRecentTransactions(
-      @PathVariable Integer accountNr, @AuthenticationPrincipal Jwt token) {
+      @PathVariable Long accountNr, @AuthenticationPrincipal Jwt token) {
     Long userId = token.getClaim(USER_ID_CLAIM);
     return ResponseEntity.ok()
         .body(bankService.getRecentTransactionsByAccountNr(accountNr, userId));
@@ -145,7 +145,7 @@ public class BankController {
   /**
    * Gets an account's total spendings in various categories from the last 30 days.
    *
-   * @param accountNr (Integer) the accounts unique number in the bank.
+   * @param accountNr (Long) the accounts unique number in the bank.
    * @param token (Jwt) the JWT token.
    * @return (Map&lt;String Double&gt;)
    */
@@ -153,7 +153,7 @@ public class BankController {
   @Operation(
       summary = "Get the sum spent in each category in the last 30 days for an account number")
   public ResponseEntity<Map<String, Double>> getRecentCategorySpendings(
-      @PathVariable Integer accountNr, @AuthenticationPrincipal Jwt token) {
+      @PathVariable Long accountNr, @AuthenticationPrincipal Jwt token) {
     Long userId = token.getClaim(USER_ID_CLAIM);
     return ResponseEntity.ok().body(bankService.getSpendingsInCategories(accountNr, userId));
   }
